+++ Assets/Mirage/Runtime/Transport/ISocket.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d97bd08ca51feda4680d85ceb95cb9e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
