+++ Assets/Mirage/Runtime/Transport/Peer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a734eca2ad00d0d4380cddbd69ea24bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
