+++ Assets/Mirage/Runtime/Transport/Transport.cs
@@ -1,156 +1,11 @@
 using System;
 using System.Collections.Generic;
-using System.Net;
-using System.Net.Sockets;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.Events;
 
 namespace Mirage
 {
-    public class UDPSocket : ISocket
-    {
-        readonly Socket socket;
-
-        public UDPSocket()
-        {
-            socket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, ProtocolType.Udp);
-        }
-
-        /// <summary>
-        /// Is message avaliable
-        /// </summary>
-        /// <returns>true if data to read</returns>
-        public bool Poll()
-        {
-            return socket.Poll(0, SelectMode.SelectRead);
-        }
-
-        public void Recieve(byte[] buffer, ref EndPoint endPoint, out int bytesReceived)
-        {
-            // todo do we need to set if null
-            endPoint = endPoint ?? new IPEndPoint(IPAddress.Any, 0);
-            bytesReceived = socket.ReceiveFrom(buffer, ref endPoint);
-        }
-
-        public void Send(EndPoint endPoint, byte[] data)
-        {
-            // todo check disconnected
-            socket.SendTo(data, (IPEndPoint)endPoint);
-        }
-    }
-    /// <summary>
-    /// Link between Mirage and the outside world...
-    /// </summary>
-    public interface ISocket
-    {
-        /// <summary>
-        /// Is message avaliable
-        /// </summary>
-        /// <returns>true if data to read</returns>
-        bool Poll();
-
-        /// <summary>
-        /// Gets next Message
-        /// <para>Should be called after Poll</para>
-        /// </summary>
-        /// <param name="data">recieved data</param>
-        void Recieve(byte[] data, ref EndPoint endPoint, out int bytesReceived);
-
-        /// <summary>
-        /// Sends to 
-        /// </summary>
-        /// <param name="data"></param>
-        void Send(EndPoint endPoint, byte[] data);
-
-        void Close();
-
-        // todo do we ever need to bind to endpoint? (rather than null)
-        void Bind(EndPoint endPoint);
-    }
-
-    public sealed class Connection
-    {
-        public readonly EndPoint EndPoint;
-
-        public Connection(EndPoint endPoint)
-        {
-            EndPoint = endPoint ?? throw new ArgumentNullException(nameof(endPoint));
-        }
-    }
-
-    /// <summary>
-    /// Controls flow of data in/out of mirage, Uses <see cref="ISocket"/>
-    /// </summary>
-    public sealed class Peer
-    {
-        // todo SendUnreliable
-        // tood SendNotify
-
-        readonly ISocket socket;
-
-        readonly Dictionary<EndPoint, Connection> connections;
-
-
-        public Peer(ISocket socket)
-        {
-            this.socket = socket ?? throw new ArgumentNullException(nameof(socket));
-        }
-
-        public void Send(Connection connection, byte[] data)
-        {
-            socket.Send(connection.EndPoint, data);
-        }
-
-        public void ReceiveLoop()
-        {
-            byte[] buffer = getBuffer();
-            while (socket.Poll())
-            {
-                //todo do we need to pass in endpoint?
-                EndPoint endPoint = null;
-                socket.Recieve(buffer, ref endPoint, out int length);
-
-                IMessageReceiver receiver = getReceiver();
-                receiver.TransportReceive(new ArraySegment<byte>(buffer, 0, length));
-            }
-        }
-
-        private IMessageReceiver getReceiver()
-        {
-            throw new NotImplementedException();
-        }
-
-        private byte[] getBuffer()
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    // todo how should we use this?
-    public sealed class PeerDebug
-    {
-        public int ReceivedBytes { get; set; }
-        public int SentBytes { get; set; }
-    }
-
-
-    /// <summary>
-    /// Creates <see cref="ISocket"/>
-    /// </summary>
-    /// <remarks>
-    /// The only job of Transport is to create a <see cref="ISocket"/> that will be used by mirage to send/recieve data.
-    /// <para>This is a MonoBehaviour so can be attached in the inspector</para>
-    /// </remarks>
-    public abstract class TransportV2 : MonoBehaviour
-    {
-        // todo rename this to Transport when finished
-        public abstract ISocket CreateSocket();
-
-        public abstract bool ClientSupported { get; }
-        public abstract bool ServerSupported { get; }
-    }
-
     [System.Obsolete("Use TransportV2, Peer, and ISocket instead", true)]
     public abstract class Transport : MonoBehaviour
     {
