+++ Assets/Mirage/Runtime/Transport/UDPSocket.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b000094fcb0965428f7a923f1770325
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
